# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BridgeWork API
  version: 0.1.0
  description: API para gestionar Projects, Tasks y TimeEntries
servers:
  - url: http://localhost:5000
paths:
  /api/projects:
    get:
      summary: List projects
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
        - name: manager_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of projects with pagination metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      per_page:
                        type: integer
    post:
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /api/projects/{project_id}:
    get:
      summary: Get a project
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    put:
      summary: Update a project
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    delete:
      summary: Delete a project
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer

  /api/tasks:
    get:
      summary: List tasks
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
        - name: project_id
          in: query
          schema:
            type: integer
        - name: assigned_to_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of tasks with pagination metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      per_page:
                        type: integer
    post:
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'

  /api/tasks/{task_id}:
    get:
      summary: Get a task
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    put:
      summary: Update a task
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    delete:
      summary: Delete a task
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer

  /api/time_entries:
    get:
      summary: List time entries
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
        - name: task_id
          in: query
          schema:
            type: integer
        - name: user_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of time entries with pagination metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/TimeEntry'
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      per_page:
                        type: integer
    post:
      summary: Create a time entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeEntryCreate'
      responses:
        '201':
          description: TimeEntry created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeEntry'

  /api/time_entries/{entry_id}:
    get:
      summary: Get a time entry
      parameters:
        - name: entry_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: TimeEntry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeEntry'
    put:
      summary: Update a time entry
      parameters:
        - name: entry_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeEntryCreate'
      responses:
        '200':
          description: TimeEntry updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeEntry'
    delete:
      summary: Delete a time entry
      parameters:
        - name: entry_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer

components:
  schemas:
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        client_id:
          type: integer
          nullable: true
        manager_id:
          type: integer
          nullable: true
        status:
          type: string
        project_type:
          type: string
        metadata_json:
          type: object
          nullable: true
        budget_hours:
          type: number
          nullable: true
        start_date:
          type: string
          format: date
          nullable: true
        end_date:
          type: string
          format: date
          nullable: true
    ProjectCreate:
      allOf:
        - $ref: '#/components/schemas/Project'
      required:
        - name

    Task:
      type: object
      properties:
        id:
          type: integer
        project_id:
          type: integer
        parent_task_id:
          type: integer
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        assigned_to_id:
          type: integer
          nullable: true
        status:
          type: string
        priority:
          type: string
        due_date:
          type: string
          format: date-time
          nullable: true
        is_external_visible:
          type: boolean
        estimated_hours:
          type: number
          nullable: true
    TaskCreate:
      allOf:
        - $ref: '#/components/schemas/Task'
      required:
        - project_id
        - title

    TimeEntry:
      type: object
      properties:
        id:
          type: integer
        task_id:
          type: integer
        user_id:
          type: integer
        date:
          type: string
          format: date
        hours:
          type: number
        description:
          type: string
          nullable: true
        is_billable:
          type: boolean
        created_at:
          type: string
          format: date-time
    TimeEntryCreate:
      allOf:
        - $ref: '#/components/schemas/TimeEntry'
      required:
        - task_id
        - user_id
        - date
        - hours